requirements.txt
orjson>=3.9.10
phonenumbers>=8.13.22
qddate>=0.1.0
pyparsing>=3.1.0
pymongo>=4.6.0
validators>=0.22.0
sqlalchemy>=2.0.0
duckdb-engine>=0.9.0
duckdb>=0.9.0
PyYAML>=6.0
typer>=0.9.0
tabulate>=0.9.0
chardet>=5.2.0
pydantic>=2.0.0
pytest>=7.4.0
xmltodict>=0.13.0
iterabledata>=0.1.0
tqdm>=4.66.0

# LLM classification dependencies
openai>=1.0.0  # For OpenAI provider and embeddings
chromadb>=0.4.0  # Vector store
requests>=2.28.0  # For OpenRouter, Perplexity, Ollama, and other HTTP-based providers

# Optional: DataHub integration
# Install with: pip install 'acryl-datahub[datahub-rest]'
# acryl-datahub[datahub-rest]>=0.12.0

# Optional: OpenMetadata integration
# Install with: pip install openmetadata-ingestion
# openmetadata-ingestion>=1.0.0

# Optional: Apache Atlas integration
# requests is already included above for LLM providers